separate hook for URL sort management
