+++ client/packages/common/src/ui/layout/tables/material-react-table/usePaginatedMaterialTable.tsx
@@ -8,7 +8,6 @@ import {
 import {
   MRT_RowData,
   MRT_RowSelectionState,
-  MRT_SortingState,
   MRT_Updater,
   MRT_PaginationState,
 } from 'material-react-table';
@@ -18,49 +17,24 @@ import { BaseTableConfig, useBaseMaterialTable } from './useBaseMaterialTable';
 interface PaginatedTableConfig<T extends MRT_RowData>
   extends BaseTableConfig<T> {
   totalCount: number;
-  initialSort?: { key: string; dir: 'asc' | 'desc' };
 }
 
+/** Use for any paginated datasets. Sort, filter and pagination must be handled externally */
 export const usePaginatedMaterialTable = <T extends MRT_RowData>({
   isLoading,
   onRowClick,
   totalCount,
-  initialSort,
   ...tableOptions
 }: PaginatedTableConfig<T>) => {
-  const {
-    updateSortQuery,
-    updatePaginationQuery,
-    queryParams: { sortBy, page, first, offset },
-  } = useUrlQueryParams({
-    initialSort,
-  });
   const t = useTranslation();
+  const {
+    updatePaginationQuery,
+    queryParams: { page, first, offset },
+  } = useUrlQueryParams();
   const [rowSelection, setRowSelection] = useState<MRT_RowSelectionState>({});
   const paginationRef = useRef<number>(0);
 
   const pagination = { page, first, offset };
-
-  const handleSortingChange = useCallback(
-    (sortUpdate: MRT_Updater<MRT_SortingState>) => {
-      if (typeof sortUpdate === 'function') {
-        // MRT can handle multiple sort fields, but for now we're only
-        // supporting one, so we take the first item of the array
-        const newSortValue = sortUpdate([
-          { id: sortBy.key, desc: !!sortBy.isDesc },
-        ])[0];
-        if (newSortValue)
-          updateSortQuery(newSortValue.id, newSortValue.desc ? 'desc' : 'asc');
-        else {
-          // For some reason, when just changing the sort direction on a field,
-          // the sortUpdate method doesn't return anything -- is this a bug in
-          // MRT?
-          updateSortQuery(sortBy.key, !sortBy.isDesc ? 'desc' : 'asc');
-        }
-      }
-    },
-    [sortBy, updateSortQuery]
-  );
 
   const handlePaginationChange = useCallback(
     (paginationUpdate: MRT_Updater<MRT_PaginationState>) => {
@@ -86,6 +60,17 @@ export const usePaginatedMaterialTable = <T extends MRT_RowData>({
   const table = useBaseMaterialTable<T>({
     isLoading,
     onRowClick,
+
+    autoResetPageIndex: false,
+    onPaginationChange: handlePaginationChange,
+    enableRowSelection: true,
+    onRowSelectionChange: setRowSelection,
+    rowCount: totalCount,
+    state: {
+      pagination: { pageIndex: pagination.page, pageSize: pagination.first },
+      showProgressBars: isLoading,
+      rowSelection,
+    },
 
     manualFiltering: true,
     manualPagination: true,
@@ -138,18 +123,6 @@ export const usePaginatedMaterialTable = <T extends MRT_RowData>({
       );
     },
 
-    onSortingChange: handleSortingChange,
-    autoResetPageIndex: false,
-    onPaginationChange: handlePaginationChange,
-    enableRowSelection: true,
-    onRowSelectionChange: setRowSelection,
-    rowCount: totalCount,
-    state: {
-      sorting: [{ id: sortBy.key, desc: !!sortBy.isDesc }],
-      pagination: { pageIndex: pagination.page, pageSize: pagination.first },
-      showProgressBars: isLoading,
-      rowSelection,
-    },
     ...tableOptions,
   });
